holo-files: Split resource.go into multiple files, tidy.
 - shorten rawResource's methods
 - remove the bogus 'if true' from StaticResource.ApplyTo()
